+++ lib/parse_args.dart
@@ -5,5 +5,9 @@
 ///
 library parse_args;
 
+export 'src/opt_def.dart';
 export 'src/opt_exception.dart';
+export 'src/opt_name.dart';
+export 'src/opt_value_count_type.dart';
+export 'src/opt_value_type.dart';
 export 'src/parse_args_base.dart';